Throw a better exception for CPF7060
The "Invalid Position" exception causes some unwanted behaviour
upstream, causing it to restart from the beginning of the journal. It
makes sense to distinguish cases like CPF7060, which are more likely
than not going to be a config error.

Also, integrate the extended message text logging into the other
exception handling, where this information would be useful.